factor out setup module
